Split and rename internal struct and files to prepare future development of the aggregator.
The number of metrics continues to increase. This commit split metrics
one per file.

Renaming 'Sampler' struct to TimeSampler (ie the sampler for
dogstatsd). This make the documentation clearer as we now have a "sampler"
concept implemented by two class: TimeSampler and CheckSampler.

Updated the README to add a diagram of the aggregator workflow.

Rename 'Metrics' to 'ContextMetrics' to avoid confusion between set of
metrics and metrics ordered by context (since this is used by both
sampler).
